+++ src/control_plane/migration/mongodb/mod.rs
@@ -1,6 +1,8 @@
 mod backup_mongodb_database;
 mod mongodb_backup_options;
 mod mongodb_connection_uri;
+mod mongodb_migration_operations;
+mod mongodb_migration_operations_options;
 mod mongodb_restore_options;
 mod mongodb_verify_target_options;
 mod restore_mongodb_database;
@@ -8,6 +10,8 @@ mod verify_mongodb_target;
 
 pub(crate) use backup_mongodb_database::backup_mongodb_database;
 pub(crate) use mongodb_backup_options::MongoDbBackupOptions;
+pub(crate) use mongodb_migration_operations::MongoDbMigrationOperations;
+pub(crate) use mongodb_migration_operations_options::MongoDbMigrationOperationsOptions;
 pub(crate) use mongodb_restore_options::MongoDbRestoreOptions;
 pub(crate) use mongodb_verify_target_options::MongoDbVerifyTargetOptions;
 pub(crate) use restore_mongodb_database::restore_mongodb_database;